experiments/original_backend/requirements.txt
fastapi
uvicorn
openai
crewai
python-dotenv
